// front-nextjs/src/client/types.gen.ts
export type UserRegister = {
    email: string;
    password: string;
    full_name?: (string | null);
};